Full Ansible control of hosts file
Using win_lineinfile is polite in that it retains hosts file entries
possibly managed outside of this role. But using said module is also
excruciatingly slow.

Thus we replace it with win_copy and jinja template the entries.

This will be a breaking change for anyone expecting to manage the hosts
file from multiple places.

Same disclaimer as in previous commit:
    Please note, no comments here about the validity of the FQDN
    list; merely moving around the existing defaults. Separate consideration
    for source data like this should be done by whomever uses the role.